create table scripts

+++ src/containers/BiletzeleWrite.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, {useEffect, useState} from "react";
 import { Auth } from "aws-amplify";
 import Form from 'react-bootstrap/Form';
 import LoaderButton from "../components/LoaderButton";
@@ -6,6 +6,8 @@ import { useAppContext } from "../libs/contextLib";
 import { useFormFields } from "../libs/hooksLib";
 import { onError } from "../libs/errorLib";
 import {useHistory, useParams} from "react-router-dom";
+import { getGames } from "../libs/utils";
+
 import "./Forms.css";
 const NO_WORDS = 5;
 
@@ -16,6 +18,7 @@ export default function BiletzeleWrite(props) {
   const wordFields = Array.from({length: noWords},(v,k)=>`word${k+1}`);
   const { userHasAuthenticated } = useAppContext();
   const [isLoading, setIsLoading] = useState(false);
+  const [playerAlreadyRegistered, setPlayerAlreadyRegistered] = useState(false);
   const [fields, handleFieldChange] = useFormFields({
     playerName: "",
     ...wordFields.reduce((map, word) => {map[word]= ""; return map;},{})
@@ -29,31 +32,50 @@ export default function BiletzeleWrite(props) {
       return false;
     return true;
   }
-  
-  function addPlayerAndWords(playerName, words){
+  function userAlreadyInGame(teams, user){
+    return teams.some(team => team.members.some(player => player.player === user.attributes.email));
+  }
+  async function addPlayerAndWords(playerName, words){
     //TODO change to getGame which will call to api query;
     const games = getGames();
-
     const game = games.find(game => game.gameId === gameId);
     const team = game.teams.find(team => team.name === teamName);
-    //TODO check if player already exists(after you add player by id)
-    team.members.push({player:"", playerName});
+    const currentUser = await Auth.currentAuthenticatedUser();
+    if(userAlreadyInGame(game.teams, currentUser)){
+      return false;
+    }
+    team.members.push({player: currentUser.attributes.email, playerName});
     game.words.push(...words);
     window.localStorage.setItem('games', JSON.stringify(games));
-  }
-
-  function getGames(){
-    return JSON.parse(window.localStorage.getItem("games"));
+    return true;
   }
 
   async function handleSubmit() {
     const words = wordFields.map(fieldName => fields[fieldName]);
-    addPlayerAndWords(fields.playerName, words);
-    history.push("/waiting-room", { params: {gameId, playerName: fields.playerName} })
+    const playerAdded = await addPlayerAndWords(fields.playerName, words);
+    if(!playerAdded) {
+      setPlayerAlreadyRegistered(true);
+    }
+    history.push("/waiting-room", {params: {gameId, playerName: fields.playerName}})
   }
+
+  useEffect(() => {
+    (async function (){
+      const games = getGames();
+      const game = games.find(game => game.gameId === gameId);
+      const currentUser = await Auth.currentAuthenticatedUser();
+      console.log(currentUser);
+      if(userAlreadyInGame(game.teams, currentUser)){
+        setPlayerAlreadyRegistered(true);
+      }
+    })();
+  }, []);
+
 
   return (
     <div className="center-form">
+      {playerAlreadyRegistered && <PlayerAlreadyRegistered/>}
+      {!playerAlreadyRegistered &&
       <Form onSubmit={handleSubmit}>
           <Form.Label>Player Name</Form.Label>
             <Form.Group controlId="playerName">
@@ -86,6 +108,11 @@ export default function BiletzeleWrite(props) {
         </LoaderButton>
 
       </Form>
+      }
     </div>
   );
+}
+
+function PlayerAlreadyRegistered(){
+  return <div> Player already registered in this game.</div>;
 }
